begin stats impl

+++ GeneralsMD/Code/GameEngine/Include/GameNetwork/GameSpy/PersistentStorageDefs.h
@@ -41,7 +41,7 @@ enum LocaleType CPP_11(: Int)
 void HandlePersistentStorageResponses(void);
 void UpdateLocalPlayerStats(void);
 
-void SetLookAtPlayer( Int id, AsciiString nick );
+void SetLookAtPlayer( int64_t id, AsciiString nick );
 void PopulatePlayerInfoWindows( AsciiString parentWindowName );
 
 #endif // __PERSISTENTSTORAGEDEFS_H__
